write and read in db

+++ nsoc/src/main/java/sensor/sensorInterface/InterfaceSensors.java
@@ -18,7 +18,7 @@ public interface InterfaceSensors {
     /**
      * @return the last data
      **/
-    float getLastValue();
+    double getLastValue();
 
 
     /**
